Changes to make it easier for other systems to extend iterate.   Expose some internal symbols as exports from the iterate/internal package.   Make special-form? a generic function so users can add their own methods (the special variable used for builtins is still there, to allow shadowing of existing special form function dispatch in the walker.)
The motivation for this change was to make iterate work
with Waters' COVER package.   The macros in that package
cannot be expanded by the walker as they introduce macrolets,
which the iterate walker cannot handle.   Instead, they
must be left unexpanded by iterate, and expanded later
by the normal macroexpansion mechanism.